Resolution and background light enhancements
Added the possibility to change the size of the OctoPiPanel window to
better with displays with resolutions other then 320x240 pixels. (Issue
#14)
It is now possible to disable the background light turn off feature by
setting the offtime to 0 in the config file. (Issue #15)

See README.md for more information on changing these settings.